Threading concept
It's only based on epoll yet (don't know how to do Winsock). There is
one thread for accepting connections from the listen socket, which are
then given to a thread with its own epoll instance. This prevents the
thundering herd issue.

The idea is to have a thread per CPU/core doing the reads, with an fd
pinned to a thread. This is not ideal theoretically because one client
can starve a thread, but Mosquitto seems based around non-blocking IO,
so I don't see that as an issue. And, it's not a regression anyway,
because that case theoretically exists already.