Only process client with changes in the main loop
Checking for keep-alives is still done, but not every loop iteration.

This fixes a considerable idle load a server has to deal with if tens of
thousands clients are only sending keep-alives.

Proof of concept, not fully tested, nor do I know if
db__message_insert() is the only place to make contexts pending.